Much progress in type improvements
