+++ packages/react-native-reanimated/src/commonTypes.ts
@@ -14,7 +14,7 @@ import type { SerializableRef, WorkletFunction } from 'react-native-worklets';
 
 import type { Maybe } from './common/types';
 import type { CSSAnimationProperties, CSSTransitionProperties } from './css';
-import type { AnyRecord } from './css/types';
+import type { AnyRecord, UnknownRecord } from './css/types';
 import type { EasingFunctionFactory } from './Easing';
 
 type LayoutAnimationOptions =
