+++ packages/react-native-reanimated/src/hook/useAnimatedProps.ts
@@ -1,38 +1,14 @@
 'use strict';
-import { SHOULD_BE_USE_WEB } from '../common';
-import type { AnimatedPropsAdapterFunction } from '../commonTypes';
-import type { DependencyList, UseAnimatedStyleInternal } from './commonTypes';
-import { useAnimatedStyle } from './useAnimatedStyle';
+
+import type { WorkletFunction } from 'react-native-worklets';
+
+import type { AnimatedPropsAdapterWorklet } from '../commonTypes';
+import type { AnimatedPropsHandle, DependencyList } from './commonTypes';
+import { useAnimatedUpdaterInternal } from './useAnimatedUpdaterInternal';
+
+const HOOK_NAME = 'useAnimatedProps';
 
 // TODO: we should make sure that when useAP is used we are not assigning styles
-
-type UseAnimatedProps = <Props extends object>(
-  updater: () => Partial<Props>,
-  dependencies?: DependencyList | null,
-  adapters?:
-    | AnimatedPropsAdapterFunction
-    | AnimatedPropsAdapterFunction[]
-    | null,
-  isAnimatedProps?: boolean
-) => Partial<Props>;
-
-function useAnimatedPropsJS<Props extends object>(
-  updater: () => Props,
-  deps?: DependencyList | null,
-  adapters?:
-    | AnimatedPropsAdapterFunction
-    | AnimatedPropsAdapterFunction[]
-    | null
-) {
-  return (useAnimatedStyle as UseAnimatedStyleInternal<Props>)(
-    updater,
-    deps,
-    adapters,
-    true
-  );
-}
-
-const useAnimatedPropsNative = useAnimatedStyle;
 
 /**
  * Lets you create an animated props object which can be animated using shared
@@ -42,12 +18,35 @@ const useAnimatedPropsNative = useAnimatedStyle;
  *   animate.
  * @param dependencies - An optional array of dependencies. Only relevant when
  *   using Reanimated without the Babel plugin on the Web.
- * @param adapters - An optional function or array of functions allowing to
- *   adopt prop naming between JS and the native side.
  * @returns An animated props object which has to be passed to `animatedProps`
  *   property of an Animated component that you want to animate.
  * @see https://docs.swmansion.com/react-native-reanimated/docs/core/useAnimatedProps
  */
-export const useAnimatedProps: UseAnimatedProps = SHOULD_BE_USE_WEB
-  ? (useAnimatedPropsJS as UseAnimatedProps)
-  : useAnimatedPropsNative;
+export function useAnimatedProps<TProps extends object>(
+  updater: WorkletFunction<[], TProps> | (() => TProps),
+  dependencies?: DependencyList | null
+): AnimatedPropsHandle<TProps>;
+
+/**
+ * @deprecated The `adapters` parameter is deprecated and will be removed in a
+ *   future version.
+ */
+export function useAnimatedProps<TProps extends object>(
+  updater: WorkletFunction<[], TProps> | (() => TProps),
+  dependencies?: DependencyList | null,
+  adapters?: AnimatedPropsAdapterWorklet | AnimatedPropsAdapterWorklet[] | null
+): AnimatedPropsHandle<TProps>;
+
+export function useAnimatedProps<TProps extends object>(
+  updater: WorkletFunction<[], TProps> | (() => TProps),
+  dependencies?: DependencyList | null,
+  adapters?: AnimatedPropsAdapterWorklet | AnimatedPropsAdapterWorklet[] | null
+): AnimatedPropsHandle<TProps> {
+  return useAnimatedUpdaterInternal(
+    HOOK_NAME,
+    updater,
+    dependencies,
+    adapters,
+    true
+  );
+}
